refactor: move Title to shared/components
