+++ src/pages/library-page.tsx
@@ -1,4 +1,3 @@
-import { Title } from "@app/title";
 import { useDeclareAppHeaderTitle } from "@app/layouts/app-header-title";
 import {
   boardPath,
@@ -18,6 +17,7 @@ import Stack from "@mui/material/Stack";
 import { EmptyState } from "@shared/components/empty-state";
 import { LoadingState } from "@shared/components/loading-state";
 import { PageContainer } from "@shared/components/page-container";
+import { Title } from "@shared/components/title";
 import { useSnackbar } from "@shared/snackbar/use-snackbar";
 import { useState } from "react";
 import { useNavigate } from "react-router";
